Move in-memory manager race test into manager_test.go
The in-memory manager's methods (Commit/RetrieveIndex) are not part of the
sealed Manager interface, so the concurrency test must access the
unexported inMemoryManager directly. Consolidate it into manager_test.go
as a package-internal (white-box) test and drop the separate file.
